requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
numpy>=1.24.0
google-generativeai>=0.3.0
python-dateutil>=2.8.0
yfinance>=0.2.30
scipy>=1.11.0
python-docx>=0.8.11
fpdf>=1.7.2
